seperated spotify functions into seperate python file, to prevent really large views.py
