+++ src/smartMirrorProject/theMirror/urls.py
@@ -1,11 +1,12 @@
 from django.urls import path
 
 from . import views
+from . import spotify
 
 urlpatterns = [
     path("", views.index, name="index"),
-    path('login/', views.login, name='login'),
-    path('callback/', views.callback, name='callback'),
-    path('get-playlist/', views.get_playlist, name='get_playlist'),
-    path('refresh-token/', views.refresh_token, name='refresh_token'),
+    path('login/', spotify.login, name='login'),
+    path('callback/', spotify.callback, name='callback'),
+    path('get-playlist/', spotify.get_playlist, name='get_playlist'),
+    path('refresh-token/', spotify.refresh_token, name='refresh_token'),
 ]
